+++ metadata.rb
@@ -2,6 +2,6 @@ name             'letsencrypt'
 maintainer       'Thijs Houtenbos'
 maintainer_email 'thoutenbos@schubergphilis.com'
 license          'All rights reserved'
-description      'Installs/Configures an acme client for Let\'s Encrypt'
+description      'Install free and trusted SSL/TLS certificates from Let\'s Encrypt'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.0'
+version          '0.1.1'