+++ recipes/default.rb
@@ -20,6 +20,7 @@
 
 chef_gem 'acme-client' do
   action :install
+  version '0.2.2'
   compile_time true
 end