+++ resources/certificate.rb
@@ -23,8 +23,11 @@ default_action :create
 
 attribute :cn,            :kind_of => String, :name_attribute => true
 
-attribute :crt,           :kind_of => String, :default => nil,   :required => true
-attribute :key,           :kind_of => String, :default => nil,   :required => true
+attribute :crt,           :kind_of => String, :default => nil
+attribute :key,           :kind_of => String, :default => nil
+
+attribute :chain,         :kind_of => String, :default => nil
+attribute :fullchain,     :kind_of => String, :default => nil
 
 attribute :owner,         :kind_of => String, :default => 'root'
 attribute :group,         :kind_of => String, :default => 'root'